pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "qalis"
version = "1.0.0"
description = "Quality Assessment Framework for LLM-Integrated Software Systems"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
keywords = [ "Artificial intelligence", "Software engineering", "Quality assessment framework", "LLM-integrated systems"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Quality Assurance",
]
dependencies = [
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "scipy>=1.11.0",
    "scikit-learn>=1.3.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
full = [
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "transformers>=4.35.0",
    "torch>=2.0.0",
    "sentence-transformers>=2.2.0",
    "openai>=1.3.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "prometheus-client>=0.18.0",
    "mlflow>=2.8.0",
    "statsmodels>=0.14.0",
    "pingouin>=0.5.3",
    "krippendorff>=0.6.0",
    "rouge-score>=0.1.2",
    "bert-score>=0.3.13",
]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
]

[project.urls]
Repository    = "https://github.com/Hussain-Md-Sahariyar/QALIS"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths    = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
markers = [
    "unit: fast unit tests (no I/O, no heavy deps)",
    "integration: tests that load data files from the repo",
    "slow: tests that instantiate transformer models",
]

[tool.coverage.run]
source = ["src/qalis"]
omit   = ["*/tests/*", "*/__pycache__/*"]

[tool.coverage.report]
show_missing = true
skip_empty   = true

[tool.black]
line-length    = 100
target-version = ["py310", "py311"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "UP"]
ignore = ["E501"]
